Refactor : 마이페이지 즐겨찾기 공용 컴포넌트 사용 (#65)
